[refactor] enum 파일 형식 수정 / 폴더 구조 생성
