--- src/main/java/com/hana/bankai/domain/account/entity/AccStatus.java
@@ -1,5 +1,0 @@
-package com.hana.bankai.domain.account.entity;
-
-public enum AccStatus {
-    ACTIVE, DELETED
-}
