+++ src/main/java/com/hana/bankai/domain/accounthistory/entity/AccountHistory.java
@@ -1,6 +1,7 @@
 package com.hana.bankai.domain.accounthistory.entity;
 
-import com.hana.bankai.global.common.enums.BankCode;
+import com.hana.bankai.domain.accounthistory.enumtype.HisType;
+import com.hana.bankai.global.common.enumtype.BankCode;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
