--- src/main/java/com/hana/bankai/domain/accounthistory/entity/HisType.java
@@ -1,5 +1,0 @@
-package com.hana.bankai.domain.accounthistory.entity;
-
-public enum HisType {
-    TRANSFER, AUTO_TRANSFER
-}
