--- src/main/java/com/hana/bankai/domain/product/entity/ProdAcc.java
@@ -1,6 +1,0 @@
-package com.hana.bankai.domain.product.entity;
-
-public enum ProdAcc {
-    SIMPLE, // 단리
-    COMPOUND // 복리
-}
