--- src/main/java/com/hana/bankai/domain/product/entity/ProdType.java
@@ -1,8 +1,0 @@
-package com.hana.bankai.domain.product.entity;
-
-public enum ProdType {
-    DEPOSIT, // 예금
-    SAVINGS, // 적금
-    LOAN, // 대출
-    DEMAND_DEPOSIT // 입출금
-}
